Made the collection/received date appearing at the top of the collection tree widget to respond to the default collection date configuration in the same way as the collection date column appearing for each specimen within the tree widget.
When defCollectionDate = 'visit_date', the common collection and
received dates will be the same as whatever populated in the visit
date field.

When defCollectionDate = 'current_date', the collection and received
dates will be the current/present date/time.

When defCollectionDate = 'none' or blank, then the user is forced
to select a date for collection and received fields.