Coordinator and receive specimens workflow improvements, bug fixes, etc
1) Display scenario specific error message when a specimen label scanned cannot
be received either because it is already received or does not exist.

2) Display scenario specific error message when specimens of the visit cannot
be received either because the visit name is invalid or specimens are not
collected or already received.

3) Hide 'Use Specimen Barcodes' checkbox when 'Use Visit Names' is clicked and
vice-versa. That is when one is clicked, the other should be hidden.

4) Display print label checkbox only at the bottom of the receive specimens form.
Do away with the checkboxes at the top.

5) Stop navigating to the nth step when specimens of multiple CPs are received.

6) Categorise the visits in the "Select Event" dropdown into 3 -
Occurred, Pending, and New.

7) Ensure no auto allocation and location dropdown are displayed in the specimen
collection page when the user does not have container read rights.

8) Remove the cross (x) button for the non-pending specimens from the collect
specimens workflow.

9) Ensure the child specimens cannot be collected when the primary specimen
is not yet received. However there is no check when the child specimens are
collected and someone changes the primary specimen to 'yet to be received'.

10) Reinitialise the specimens in the collection tree whenever the user switches
between visits.

11) Add checks to ensure coordinators cannot delete the participants, visits
that have one or more received specimens underneath them.

12) Rename the "Submit & Print" button to "Submit & Print Manifest".

13) Added options in the receive workflow to allow users print the labels
of primary, derived and aliquot specimens.

14) Many other minor stuff along the way.