Speed up the staged participants lookup by MRN.
The strategy employed is to do the lookup in 2 phases:
a) Find a matching participant ID in the MRN table based on site name and MRN.
b) Use the matched participant IDs to retrieve the participant details.

This is much faster than joining 2 large tables of 7 million + entries.

The lookup results are appearing well within 3 seconds.