OPSMN-5756: Added a null guard in the print rule to ensure there are no NPEs when user groups are edited.
Whenever a user group is edited / deleted, an event is broadcasted.
The broadcasted event is catched by the interested parties to
carry out necessary actions. Print rule is one of the interested
party. Its job is to ensure the print rule is not referring to the
deleted user groups or users removed from the user group etc.
During this event handling, the print rule was crashing with NPE.