Coordinator workflow feedback changes.
*) Changed the label of "Add Participant" to "Add/Select Participant"
Applicable only when the button/menu option navigates users to the
RDE workflow. In all other cases, it will be "Add Participant".

*) Changed the label "Select Event" to "Select Event (timepoint)"
on the participant specimen collections step of the RDE/Coordinator
workflow.

*) Added specimen description column to the bulk add/edit events table.

*) Displayed the "User" column as "Receiver" in the receive specimens
table. The received event form is modified on fly to change the
label from "User" to "Receiver".

*) Added a setting to specify the specimens that need to be printed
on receiving primary specimens. The setting value is comma-separated
list of specimen lineages. For example - primary,derived,aliquot

*) In the receive specimens UI, by default, "Scan visit names" is
selected.

*) Updated the workflow setting to allow CP desginers specify
the default value for specimen collection status based on their lineage.

"defCollectionStatus" : {
  "New": "Pending",
  "Derived": "Collected",
  "Aliquot": "Collected"
}

When the value of setting is primitive string, then it is assumed to
be the default value for all the lineages.

"defCollectionStatus": "Pending"

translates to

"defCollectionStatus" : {
  "New": "Pending",
  "Derived": "Pending",
  "Aliquot": "Pending"
}