OPSMN-5807: The UI is disabled when a form is submitted to the backend / API.
a) Implemented a general purpose mask component which can be used to cover/mask the UI. This gives UI disabled feeling. All elements of the UI can be seen but cannot be touched.

b) Updated the HttpClient to emit events when backend calls are made. This allows other components of the UI to listen for the events and take appropriate actions.