OPSMN-5862: There was a race condition between loading of the translation files and UI rendering. When the translation files are loaded before UI rendering, everything works fine without any glitches. However, the problem arises when UI is rendered before the translation files are loaded. This is when certain widgets like side menu stop working.
The UI initiates the loading of translation files from the server and
continues with its other tasks. Many a times, these files are loaded
before the UI starts rendering... To make this work all the times, added
a barrier to ensure the rendering does not start before the translation
files are loaded.