OPSMN-5856: Added a new preprint setting - On Shipment Receive
With this setting enabled, whenever a primary specimen is received,
its child specimens are automatically created in the pending state
and their labels are printed.

Added a global setting to allow users filter the allowed received
qualities on which the pre-printing should be triggered.
By default, the setting value is blank, which is to treat
any value as the allowed received quality.