OPSMN-5920, OPSMN-5918: Sanitised order name for use in the report filenames.
When a order is created, a CSV/PDF report is created. The report filename
is based on order name. If the order name contains special characters like
forward slash (/), colon (:) etc, then the API encounters file system error.
The forward slash (/) is used as a path separator on Linux/Mac. The colon
is used in drive names on Windows.

Therefore sanitised the order name before using it in the report filenames.
The idea is replace all non-word characters (excluding dot) with an
underscore. Two or more successive underscores are replaced by a single
underscore.

For example - Test DP Jun 15, 2022 18:36:59 is transformed to
Test_DP_Jun_15_2022_18_36_59 before using it in the report filename.

Similarly fixed the container report APIs and UI to replace the special
chars with underscore.