OPSMN-5886: Propagated the database errors encountered during form creation or edit op to the users. This will help users to know the cause of the error either by searching the Wiki articles or connecting with the support staff.
The error message "Error saving container" was not of much help. Neither
the users, support staff or the developers understood the cause resulting
in exchange of emails. This meant lower ESG scores...

With this fix, the error displayed to the users is -

Form Error: Error saving container: StatementCallback; bad SQL grammar
[ALTER TABLE DE_E_11551 ADD (DE_A_86 VARCHAR(255))]; nested exception
is java.sql.SQLSyntaxErrorException: Row size too large. The maximum
row size for the used table type, not counting BLOBs, is 65535.
This includes storage overhead, check the manual. You have to change
some columns to TEXT or BLOBs