Vue JS: Initialise the multiselect dropdown options using the selected values...
When a form is loaded, the dropdown options are not loaded. The options are
loaded only when the dropdown is opened. Further, the selected values/options
are not displayed if they are present in the dropdown options list. This is
undesirable scenario.

To overcome this problem, the dropdown options list is initialised using
the selected values.