Vue JS: Enhanced the overview widget to display fields based on criteria.
For example, on the container overview page, it doesn't make sense to
display the location field for the parent containers. Same way, there is
no point in displaying dimensions field for a dimensionless container.

Enhanced overview widget to display or hide the field based on the
conditions specified in the view metadata.