OPSMN-5954: Show at max 500 command files on the UI.
The order in which these 500 files are selected is left to the operating system.
However, the first 500 files picked by the operating system are sorted in
reverse order of their creation/modification.

This measure is implemented to ensure the server is not drowned by large
number of print files present on the test instances.