DE library containing the following fixes
OPSMN-5968: Strict deserialisation of form XML.
Do not deserialise any object that are not required to be present in the form
metadata. This is done by creating a whitelist of the allowed objects and
interfaces.

OPSMN-5969: Ensure there is no path traversal in the options
file specified in the input XML.
This is to prevent local file inclusion attacks.