OPSMN-5970: Configurable limit and asynchronous bulk specimens update via UI.
a) Added a configuration setting to specify the maximum number of specimens
that can be updated at a time using the UI or direct API calls.

b) Turned the bulk specimens update API into asynchronous API. If the
update finishes within 30 seconds, the API invoker receives the response
instantaneously. Otherwise, an email notification is sent on completion
or failure of the API.