OPSMN-5986: Fixed the Liquibase changeset to drop any existing foreign keys on the older user institute ID column.
The changeset was working fine until we upgraded to Liquibase 4.8.0.
The newer Liquibase requires the changeset to specify the name of
the table on which FK checks needs to be done. This is in contrast
to earlier versions that only required the FK name.

Earlier:
if (foreign key 'xyz' exists) {
  drop foreign key 'xyz'
}

Newer:
if (foreign key 'xyz' exists on table 'abc') {
  drop foreign key 'xyz' on table 'abc'
}