Upgraded Spring, Hibernate, and Velocity libraries. First cut.
Major changes because of deprecation or removal of Criteria API
in the Hibernate framework. The older Criteria API is replaced
by CriteriaBuilder.

Added our own Criteria interface to absorb the future shocks
of changes in the CriteriaBuilder and its associated APIs.
Also, the interface is more closer to older Criteria API in usage.