OPSMN-6083: Order Add/Edit - Automatic turn on/off dispose specimen checkbox
The dispose specimen checkbox is automatically checked on when either of the
following conditions are met:
a) Specimen is stored in holding tanks.
b) Specimen available quantity is not specified or 0.
c) Distribution quantity is the same as available quantity.

The dispose specimen checkbox is automatically checked off when all of the
following conditions are met:
a) Specimen is not stored in holding tanks.
b) Specimen available quantity is greater than 0.
c) Distribution quantity is less than the available quantity.