OPSMN-6073: Specimen requirement update API.
It is possible to change any existing specimen's requirement via bulk import
or API call.

When a specimen's requirement is updated, all its children's requirements
are nullified. The updated requirement lineage and type should match with
that of the specimen.

If a child specimen's requirement is updated then the new requirement's
parent should be the same as the child's parent requirement.