System user ($system) does not have any institute or auth domain.
Audit logs on the UI didn't reflect the changes (creation/modification)
made by the system user (as in HL7, REDCap, OC), as the query was
expecting every user to have an institute and auth domain.

Fixed the query to obtain the revisions by using left join on institutes
and auth domains tables.