OPSMN-6093: Added a couple of configuration settings to allow users input specimen external IDs at the time of receiving the shipment.
a) Add External ID: When enabled, a couple of columns - External ID: Name,
and External ID: Value are displayed on the shipment receive page.
These columns/fields allow users to specify the new external ID for
the specimens being received. These fields will be rendered as optional
fields. However, when the value of one field is specified then the other
field will be mandatory.

b) External ID Name: Default name for the external ID added at the
time of receiving the shipment. When this setting is specified, only
one column - External ID: Value (optional) is displayed on the shipment
receive page.

The newly specified external IDs are added to the specimens' existing
external IDs list.

If the name of new external ID matches with any of the existing IDs,
then the value of matched external ID is overwritten by the new value
specified by user during shipment receive.