OPSMN-6095, OPSMN-6096: Cart improvements.
1) Carts can be shared with users of multiple institutes.

2) Added carts digest. The digest includes the count of specimens added
by different cart users since the last digest run.

The digest email is sent to all the cart collaborators - owner, shared
users, and shared group users.

No digest is sent when there are no specimens added to the cart since
last run.

Digest can be enabled at the individual cart level.

Digest frequency is configurable at the system level via the job settings.
The frequency can be as high as every minute (minutely) or as low as
once in a month (monthly).