Automated freezer report improvements.
1. Generated the report per automated freezer configured in the system.
This is slightly different than the earlier approach in which a single
report was generated for all the freezers.

Given that there will be only one automated freezer, this should not
make much difference.

2. The report is dispatched to freezer site administrators.
Again this is different from the earlier approach in which the report
was sent only to the system administrators.

3. Added collection protocol short title in the generated report.