OPSMN-6158: Unlink eMPI and other participant identifiers from the deleted registration.
Majority of times, when a registration is deleted the corresponding participant object/record
is deleted as well. However, if the participant is associated with multiple registrations then
deletion of registration will not result in deletion of participant. In such cases, the
search indexer does not receive the participant delete event As a result, it does not get
a chance to severe the link between the participant eMPI/other identifiers with the
deleted registration.

Fixed the registration search indexer to unlink eMPI and other identifiers from the
deleted registration.