OPSMN-6164: Clear out the receiving site and  requester when receiving institute is updated.
Receiving institute is a derived field in the order object model.  That is, the
receiving institute is not persisted in the database. Instead the requester institute
is used to populate it. So if a user changes only the receiving institute without
changing the requester then no change is effected.

To solve the problem, receiving site and requester fields are cleared when the
receiving institute is changed. This forces the user to select the requester
from the new institute.