Fixed the container restrictions checker.
When the parent container without any CPs is updated, a check is made
to ensure that all the child containers (at any level) still have the
same set of CPs that run at the parent container site. This check is
performed using a complicated query. The query was broken when moving
to a new Hibernate Criteria based API. Fixed the same query that fits
in the semantics of the new API.