OPSMN-8489: Used materialised table to query the registration/participant details.
Earlier a view that joined the registration and participant tables was used.
This was creating performance bottlenecks when the view was used twice or thrice
in the same query.

To solve the problem created a materialised table that contains the join of
registration and participant tables. This materialised table is kept in sync
with the source tables via triggers.