OPSMN-6125: Added supporting APIs / attributes for participant emails
a) Added email sender group attribute at the CP level.
The sender group defines the list of users who can send emails to the participants.

b) Added user group API to check whether the current signed-in user is
a member of the requested group or not.

c) Added an email API to obtain the content of the email that would be sent given
the input template and template parameters. The content will be used for displaying
the email preview.