OPSMN-6228: Correctly handle sub-form row deletion.
The sub-form row delete was implemented using row indices.
Row index changes when rows are deleted. For example, when
row #1 is deleted, the row #2 becomes row #1 and like that.

To handle deletion, every row is assigned an unique ID (UID).
The UID is used to identify and delete the sub-form rows.