OPSMN-6285: Find free places to store boxes/containers.
a) Users can specify criteria for selecting the parent containers
with free locations to store the containers in hand.

b) Containers that satisfy the users criteria are displayed in
list view sorted by free locations in descending order.

c) Users can select one or more parent containers to store the
child containers.

d) Child containers can be existing or brand new. The existing
containers will be transferred to the new parent container.
New containers will be created on the fly.