OPSMN-6311: The problem was - the input CSV consent values were not serialised to the Java interaction object fields that are used when invoking the APIs.
This was because of marking the statement, code, and response fields with
JsonIgnore tag to ensure these are not present in the consent response API
payload. However, this also had an adverse impact on the bulk import.

The fix is to remove the JsonIgnore annotation on these fields - code,
statement, and response, and mark the interaction object to ignore the
null fields when serialising to JSON.