OPSMN-6340: When query results are exported, a file ID is sent in response to the client/user. This file ID can used to download the exported CSV file using the query API. Mapping the file ID to the exported CSV file path and delivering the requested file is query API's responsibility.
When query results are exported using scheduled jobs, the resultant CSV file
path is stored in the job run. This allows the jobs API to deliver the
required file whenever requested. The jobs API doesn't distinguish or doesn't
have to care about who created the file. Simple.

The job file name was conflicting with the temporary file created by the query
API to export the data. As a result, when the query API deleted all the temporary
files created by it, it deleted the result file as well.

To fix the problem, the result and temporary file names are set to different values.
For example, the result file will be scheduled_query_<query_id>_<run_id>.csv.zip,
temporary file will be <result_file>.zip