Fixed the specimen and container APIs to spit out the incorrect specimen class/type in the error messages.
Earlier: Specimen type is invalid

New: Specimen type {0} is invalid.

Note: When the API was written in 2014/15, it was in context of a single specimen.
Over time, we've been using the API in many ways than imagined. The same APIs
are used by the workflow module to create/update many specimens in one go.
The earlier message doesn't help users identify which of the many specimens is
using invalid class/type.