Fixed the Vue based PV dropdown widget to cache the PV query results.
For example, when a table form like create specimens task contain multiple
specimen rows, each with the same PV - tube type and value, the dropdown
resulted in issuing as many queries as the number of specimens/rows in the
table form. This is inefficient, freezes the browser...

The fix is to cache query result by PV attribute + query params and use
it whenever a similar requirement arises. The caching is done per form.
That means, the cache vanishes when the form is destroyed.