Table form improvements
a) Support for tree layout as in the specimen collection tree

b) Ability to hide/show fields of the row independent of the other rows.
That is, a field f could be displayed for one row but hidden for another.
For example, collection container displayed for all the primary specimen
rows but hidden for the children. Storage container displayed for all
aliquots but hidden for others.

This is done using showCellWhen. This is different from showWhen

showWhen is used to control whether a field should be displayed or
not for all the rows. Example - whether holding location should be
displayed or not based on the selected DP. It will be same for all
the rows.

showCellWhen is used to control whether a field of a given row should
be displayed or not.