OPSMN-6281: Participant messaging improvements.
a) Added boolean settings to indicate whether the participant has
opted-in for receiving messages via emails or SMS or both.

b) By default, the existing participants with email addresses are
assumed to have opted-in for receiving messages via emails.

c) Bulk import / export for the new opt-in settings.

d) Support to query the participant opt-in settings.