NUI-0112: Some columns of the specimen tree were not displayed because the conditional logic specified in the dictionary evaluated to false.
This behaviour is different from the older specimen tree.

To accommodate both the new and old behaviour, when the condition to display
evaluates to false, the table column is displayed without the value. This means
the tree will end up with some rows having value for the column
and no value for the others.

For example:
{
  name: specimen.receivedEvent.time
  caption: Received Time
  showWhen: 'specimen.lineage == 'New'
}

This will display the received time for only primary specimens.
Other specimens of the tree will have blank or '-' in the received time column.