OPSMN-6465: Allowed only one audit report to be generated at any point in time across any number of nodes in the cluster.
Implemented general purpose locking mechanism that can be used for any future/other
purposes.

The thread that is generating the audit report, first acquires the lock. The thread
relinquishes the lock only after the report is generated or any error is encountered.
When another thread tries to generate the report, it has to first acquire the lock.
If it fails in acquiring the lock, it will give up with the error:
"Audit report generation is already in progress. Only one audit report can be
generated at a time. Please retry after some time. (AUDIT_REPORT_GEN_IN_PROGRESS)"