UI component enhancements
1. Added "More" menu plugin hooks on the participant, visit, and
specimen overview pages.

2. Relaid the form record overview UI. Moved the action buttons from
the table rows to the record overview card.

3. Added plugin hooks in the form and specimen event record views.

4. Extracted the display value logic from the span component into a
utility method. Refactored the span component to depend on the utility
method for displaying various field values.

The utility method can be used to capture any field's display value
including the subforms, signature and file upload widgets/components.

5. Improved the menu component. The menu dropdown button is displayed
only if the menu has more than 1 option. The  menu dropdown is
displayed as an "option button" if the menu has a single option.
The menu dropdown is hidden when there are no menu options.