OPSMN-6603: An email notification is sent whenever the shipment request status is modified.
The email notification is sent to the shipment notif users, if any, shipment
sender, shipment requester.

The email clearly highlights the old status and new status.

Allowed users to filter the shipments list based on the request status.

It is obvious the next request from users would be to see the request
status on the list view. Therefore the request status is also shown in the
list view along with the shipment status.