OPSMN-6621: Modified form files API to ensure only the authorised users can download the files and signatures. Knowing file ID alone is no longer sufficient to download the files...
Added a migration script to crawl through all the existing forms and their records, store their
file IDs in the dyextn_form_files table along with the form records (IDs) owning them:

(form_id, record_id, file_id)

When a file with certain ID is requested for download, the API searches for the relevant
record using file ID as the key. The download request is honored if the record is found
and the user has read access on the record (form_id, record_id). Otherwise, the request
is rejected.