OPSMN-6626: Fixed automated code scanning issues (potentially security threats)
1. Created temp files inside $OS_DATA/tmp directory instead of one the specified in
java.io.tmpdir. This ensures temporary files can be accessed by only those VM users
who have access to OpenSpecimen data.

2. Measures to prevent ZIP bomb attacks. This is done by ensuring, the number of
files/entries in the inflated ZIP are not more than 100 and cumulative inflated size
is not more than 1GB.

3. Replaced the regular expression that take polynomial time with the simpler & restricted
string manipulation logic.

4. Used SecureRandom to generate pseudo-random numbers instead of Random. Even Random
was safe because it was invoked as many times as the number of digits in the OTP.

5. Ensured all the stack traces are printed in OpenSpecimen log files instead of Tomcat's
console/log file.