OPSMN-6920: Fixed multiselect dropdown render bugs
a) An hidden element was added to the bottom of the page that was causing
the page to scroll when the dropdown was opened.

Fixed the CSS of the hidden element so that it is rendered in the same
container (div) as the dropdown.

b) Used the same array used by the dropdown to update it with the
selected options.