Fixed the error code/message that is sent in response when the query is aborted after 55 seconds limit.
The recent Spring data access API and MySQL driver combination results in an unfathomable
exception - UndeclaredThrowableException. OpenSpecimen rightly considers this to be some
kind of an unhandled server error and tries to handle it...

Eventually the users see an error that they should be least bothered about.

The fix is to traverse the exception tree and check whether "UndeclaredThrowableException"
was due to SQLTimeoutException. If yes, return QUERY_TIMEOUT error in the API response.