Rewrite of transaction scoped cache.
The earlier transactional cache had a bug in that the cache was not cleared
when a database error (like unique constraint, data too long etc) was encountered.

Earlier transactional cache relied on Hibernate interceptor for notification
of transaction start and end. The transaction end notification was not raised
when database error was encountered. As a result, the transactional cache didn't
get cache to clear...

Rewrote the entire cache using Spring's transaction synchronisation APIs.
This resulted in some interface changes and therefore the consuming code.