OPSMN-7138: Table based data entry forms: Implemented Zebra Striping.
All even rows of the table will have a light grayish background and the odd rows
will have the regular white background.

(Personally, I do not like the Zebra striping idea. I feel it adds more visual
noise than helping users do the data entry.

How will user distinguish between 2 same coloured rows if there was a small
distraction or user has to move their eyes back to the screen?

The right way is to use sticky left columns that display primary attributes of
the row and use it as a reference to decide whether data is being entered in
the right row or not)