OPSMN-7140: Added config setting to limit the container size/dimension.
Previously, users were able to create containers with arbitrary dimensions.
In production, we have seen cases like containers or boxes with dimensions
1800 rows x 1 column, which in all likelihood is impractical and unintended.

This leads to performance issues. When allocating the next available position,
the app needs to scan the entire container map and load all occupants to find
the next vacant position. This process would significantly increase the
application memory footprint.

While we intend to optimise the allocation mechansim, until we do so, the fix
is to ensure the containers are of restricted size. By default, containers
are restricted to 225 positions.